decoder: refactor and simplify
The decode_token function returning both an (potentially) valdated and
an unvalidated token decoding as well as the printing format makes the
function a somewhat odd shape. Split the format recognition out as it's
unrelated to decoding. Also only return one result for an (optional)
verified token as practically the unvalidate token data was only used
iff the token also successfully validated.

While adjusting the tests also simply unwrap the results rather then
asserting `.is_ok()`, which has the benefits of printing the Err on
failure.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>